+++ lib/ayeaye/connectors/fake.py
@@ -13,24 +13,12 @@ class FakeDataConnector(DataConnector):
     }
     preserve_callables = ["quantum_factory"]
 
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
     @property
     def engine_params(self):
         """
         :returns: (dict) of parameters needed to connect to engine without secrets.
         """
         return {"engine_url": "fake://example.com/abc"}
-
-    def connect(self):
-        pass
-
-    def __len__(self):
-        raise NotImplementedError("TODO")
-
-    def __getitem__(self, key):
-        raise NotImplementedError("TODO")
 
     @property
     def data(self):